feat: add global component generator qr code with base64 string
